refactor: update algolia routes
