+++ src/v2/Apps/Algolia/routes/AlgoliaHome.tsx
@@ -1,0 +1,9 @@
+import { Box, Text } from "@artsy/palette"
+
+export const AlgoliaHome = () => {
+  return (
+    <Box my={4}>
+      <Text>AlgoliaHome</Text>
+    </Box>
+  )
+}
