+++ src/v2/Apps/Algolia/routes/AlgoliaResults/Components/AlgoliaIndices.tsx
