+++ src/v2/Apps/Algolia/routes/AlgoliaResults/Components/__tests__/AlgoliaIndices.jest.tsx
